Catching Errors in Async Functions
we’ve used try/catch blocks to catch errors in our async/await functions, but they make our code look messy. To fix this issue, we’ll put the try/catch block into a higher function and then put the async/await functions of that function. The start of that refactoring looks like this