Creating an Error Class
The super() function only takes message as an argument because that’s what Error takes natively. Then, we add a statusCode property and a status that derives from statusCode. The isOperational property will come in handy later. Finally, the captureStackTrace line prevents this class from showing up in the stack trace, which is part of the console log that shows where in code the error occurred.